+++ relay-server/src/utils/sizes.rs
@@ -25,6 +25,7 @@ pub fn check_envelope_size_limits(config: &Config, envelope: &Envelope) -> bool 
             ItemType::Event
             | ItemType::Transaction
             | ItemType::Security
+            | ItemType::ReplayEvent
             | ItemType::RawSecurity
             | ItemType::FormData => event_size += item.len(),
             ItemType::Attachment | ItemType::UnrealReport | ItemType::ReplayRecording => {