+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 **Bug Fixes**:
 
-- Parse custom units with length < 15 without crashing. ([#1311](https://github.com/getsentry/relay/pull/1311))
+- Parse custom units with length < 15 without crashing. ([#1312](https://github.com/getsentry/relay/pull/1312))
 
 ## 22.6.0